Break apart Rom class, add migrate command
